fixed warshall_floyd

+++ library.cpp
@@ -3483,7 +3483,7 @@ vector<vector<long long>> warshall(vector<vector<edge<T>>> &g){
   int n = (int)g.size();
   vector<vector<long long>> d(n, vector<long long> (n, LINF));
   for(int i = 0; i < n; i++) d[i][i] = 0;
-  for(int i = 0; i < n; i++) for(edge<T> e : g[i]) d[i][e.to] = min(d[i][e.to], e.cost);
+  for(int i = 0; i < n; i++) for(edge<T> e : g[i]) d[i][e.to] = min(d[i][e.to], (long long)e.cost);
   for(int k = 0; k < n; k++) for(int i = 0; i < n; i++) for(int j = 0; j < n; j++){
     if(d[i][k] != LINF && d[k][j] != LINF){
       d[i][j] = min(d[i][j], d[i][k] + d[k][j]);
